requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
Jinja2==3.1.2
Pillow==10.0.0
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
gunicorn==21.2.0
python-dotenv==1.0.0
Markdown==3.5.1
requests==2.31.0